+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 
 * Add rule LOG013 that detects mismatches between named ``%``-style formatting placeholders and keys in dict argument.
 
+* Add rule LOG013 that detects ``exc_info=True`` outside of exception handlers.
+
 1.3.1 (2023-09-17)
 ------------------